Fixed more unsigned to signed type comparison warnings, changed several variables back to type int where size_t is not required. Fixed a bug where the user could scroll up out of bounds on the read loop sorted by date due to comparing an unsigned type sub 1 compared greater to or equal than 0.

+++ src/main.h
@@ -52,7 +52,7 @@ struct Categories {
 
 struct ScrollCursorSimple {
 	size_t displayed;
-	size_t select_idx;
+	int select_idx;
 	int cur_y;
 };
